+++ R/vntrs-package.R
@@ -1,7 +1,5 @@
 #' @aliases vntrs-package
+#' @useDynLib vntrs, .registration = TRUE
+#' @importFrom Rcpp evalCpp
 #' @keywords internal
 "_PACKAGE"
-
-## usethis namespace: start
-## usethis namespace: end
-NULL