+++ NEWS.md
@@ -1,3 +1,11 @@
+# vntrs 0.2.0
+
+* Added optional bound constraints to the trust-region optimizer and made them available through the `lower` and `upper` arguments.
+  
+* Improved documentation.
+
+* Implemented algorithm in Rcpp for speed increase.
+
 # vntrs 0.1.1
 
 * Small bug fixes.